# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Microblogging API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PostContentFormat = typeof PostContentFormat[keyof typeof PostContentFormat];


export const PostContentFormat = {
  plain: 'plain',
  html: 'html',
} as const;

export interface Post {
  id: number;
  authorId: string;
  authorName: string;
  authorImageUrl?: string | null;
  content: string;
  contentFormat: PostContentFormat;
  commentCount: number;
  createdAt: string;
}

export interface PostsPage {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}

export interface Comment {
  id: number;
  postId: number;
  authorId: string;
  authorName: string;
  authorImageUrl?: string | null;
  content: string;
  createdAt: string;
}

export interface PostWithComments {
  post: Post;
  comments: Comment[];
}

export type CreatePostBodyContentFormat = typeof CreatePostBodyContentFormat[keyof typeof CreatePostBodyContentFormat];


export const CreatePostBodyContentFormat = {
  plain: 'plain',
  html: 'html',
} as const;

export interface CreatePostBody {
  /** @maxLength 40000 */
  content: string;
  contentFormat: CreatePostBodyContentFormat;
}

export type UpdatePostBodyContentFormat = typeof UpdatePostBodyContentFormat[keyof typeof UpdatePostBodyContentFormat];


export const UpdatePostBodyContentFormat = {
  plain: 'plain',
  html: 'html',
} as const;

export interface UpdatePostBody {
  /** @maxLength 40000 */
  content: string;
  contentFormat: UpdatePostBodyContentFormat;
}

export interface CreateCommentBody {
  /** @maxLength 500 */
  content: string;
}

export interface UpdateCommentBody {
  /** @maxLength 500 */
  content: string;
}

export interface UserProfile {
  id: string;
  name: string;
  imageUrl?: string | null;
  postCount: number;
}

export interface FeedStats {
  totalPosts: number;
  totalComments: number;
}

export interface UploadedMedia {
  url: string;
  mimeType: string;
  width?: number | null;
  height?: number | null;
}

export type ListPostsParams = {
page?: number;
limit?: number;
};

export type GetPostsByUserParams = {
page?: number;
limit?: number;
};

export type UploadMediaBody = {
  file: Blob;
};