# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Microblogging API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: posts
    description: Post operations
  - name: comments
    description: Comment operations
  - name: users
    description: User operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List all posts (paginated)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Unauthorized

  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a single post with its comments
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post with comments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostWithComments"
        "404":
          description: Not found
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /posts/user/{userId}:
    get:
      operationId: getPostsByUser
      tags: [posts]
      summary: Get all posts by a specific user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: User's posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostsPage"

  /posts/{postId}/comments:
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a post
      parameters:
        - in: path
          name: postId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized

  /comments/{id}:
    patch:
      operationId: updateComment
      tags: [comments]
      summary: Update a comment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommentBody"
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (owner only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized

  /feed/stats:
    get:
      operationId: getFeedStats
      tags: [posts]
      summary: Get feed statistics (total posts, active users today)
      responses:
        "200":
          description: Feed statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedStats"

  /media:
    post:
      operationId: uploadMedia
      tags: [posts]
      summary: Upload a media asset for a post (owner only)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: Uploaded media metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadedMedia"
        "400":
          description: Invalid upload
        "401":
          description: Unauthorized
        "403":
          description: Forbidden

  /media/{fileName}:
    get:
      operationId: getMedia
      tags: [posts]
      summary: Fetch an uploaded media asset
      parameters:
        - in: path
          name: fileName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Media content
        "404":
          description: Not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Post:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorImageUrl:
          type: string
          nullable: true
        content:
          type: string
        contentFormat:
          type: string
          enum: [plain, html]
        commentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - authorId
        - authorName
        - content
        - contentFormat
        - commentCount
        - createdAt

    PostsPage:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - posts
        - total
        - page
        - limit

    Comment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorImageUrl:
          type: string
          nullable: true
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - postId
        - authorId
        - authorName
        - content
        - createdAt

    PostWithComments:
      type: object
      properties:
        post:
          $ref: "#/components/schemas/Post"
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required:
        - post
        - comments

    CreatePostBody:
      type: object
      properties:
        content:
          type: string
          maxLength: 40000
        contentFormat:
          type: string
          enum: [plain, html]
      required:
        - content
        - contentFormat

    UpdatePostBody:
      type: object
      properties:
        content:
          type: string
          maxLength: 40000
        contentFormat:
          type: string
          enum: [plain, html]
      required:
        - content
        - contentFormat

    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
          maxLength: 500
      required:
        - content

    UpdateCommentBody:
      type: object
      properties:
        content:
          type: string
          maxLength: 500
      required:
        - content

    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        imageUrl:
          type: string
          nullable: true
        postCount:
          type: integer
      required:
        - id
        - name
        - postCount

    FeedStats:
      type: object
      properties:
        totalPosts:
          type: integer
        totalComments:
          type: integer
      required:
        - totalPosts
        - totalComments

    UploadedMedia:
      type: object
      properties:
        url:
          type: string
        mimeType:
          type: string
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
      required:
        - url
        - mimeType