# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * Microblogging API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary List all posts (paginated)
 */
export const listPostsQueryPageDefault = 1;
export const listPostsQueryLimitDefault = 20;

export const ListPostsQueryParams = zod.object({
  "page": zod.coerce.number().default(listPostsQueryPageDefault),
  "limit": zod.coerce.number().default(listPostsQueryLimitDefault)
})

export const ListPostsResponse = zod.object({
  "posts": zod.array(zod.object({
  "id": zod.number(),
  "authorId": zod.string(),
  "authorName": zod.string(),
  "authorImageUrl": zod.string().nullish(),
  "content": zod.string(),
  "contentFormat": zod.enum(['plain', 'html']),
  "commentCount": zod.number(),
  "createdAt": zod.coerce.date()
})),
  "total": zod.number(),
  "page": zod.number(),
  "limit": zod.number()
})


/**
 * @summary Create a new post
 */
export const createPostBodyContentMax = 40000;



export const CreatePostBody = zod.object({
  "content": zod.string().max(createPostBodyContentMax),
  "contentFormat": zod.enum(['plain', 'html'])
})


/**
 * @summary Get a single post with its comments
 */
export const GetPostParams = zod.object({
  "id": zod.coerce.number()
})

export const GetPostResponse = zod.object({
  "post": zod.object({
  "id": zod.number(),
  "authorId": zod.string(),
  "authorName": zod.string(),
  "authorImageUrl": zod.string().nullish(),
  "content": zod.string(),
  "contentFormat": zod.enum(['plain', 'html']),
  "commentCount": zod.number(),
  "createdAt": zod.coerce.date()
}),
  "comments": zod.array(zod.object({
  "id": zod.number(),
  "postId": zod.number(),
  "authorId": zod.string(),
  "authorName": zod.string(),
  "authorImageUrl": zod.string().nullish(),
  "content": zod.string(),
  "createdAt": zod.coerce.date()
}))
})


/**
 * @summary Update a post (owner only)
 */
export const UpdatePostParams = zod.object({
  "id": zod.coerce.number()
})

export const updatePostBodyContentMax = 40000;



export const UpdatePostBody = zod.object({
  "content": zod.string().max(updatePostBodyContentMax),
  "contentFormat": zod.enum(['plain', 'html'])
})

export const UpdatePostResponse = zod.object({
  "id": zod.number(),
  "authorId": zod.string(),
  "authorName": zod.string(),
  "authorImageUrl": zod.string().nullish(),
  "content": zod.string(),
  "contentFormat": zod.enum(['plain', 'html']),
  "commentCount": zod.number(),
  "createdAt": zod.coerce.date()
})


/**
 * @summary Delete a post (owner only)
 */
export const DeletePostParams = zod.object({
  "id": zod.coerce.number()
})


/**
 * @summary Get all posts by a specific user
 */
export const GetPostsByUserParams = zod.object({
  "userId": zod.coerce.string()
})

export const getPostsByUserQueryPageDefault = 1;
export const getPostsByUserQueryLimitDefault = 20;

export const GetPostsByUserQueryParams = zod.object({
  "page": zod.coerce.number().default(getPostsByUserQueryPageDefault),
  "limit": zod.coerce.number().default(getPostsByUserQueryLimitDefault)
})

export const GetPostsByUserResponse = zod.object({
  "posts": zod.array(zod.object({
  "id": zod.number(),
  "authorId": zod.string(),
  "authorName": zod.string(),
  "authorImageUrl": zod.string().nullish(),
  "content": zod.string(),
  "contentFormat": zod.enum(['plain', 'html']),
  "commentCount": zod.number(),
  "createdAt": zod.coerce.date()
})),
  "total": zod.number(),
  "page": zod.number(),
  "limit": zod.number()
})


/**
 * @summary Add a comment to a post
 */
export const CreateCommentParams = zod.object({
  "postId": zod.coerce.number()
})

export const createCommentBodyContentMax = 500;



export const CreateCommentBody = zod.object({
  "content": zod.string().max(createCommentBodyContentMax)
})


/**
 * @summary Update a comment
 */
export const UpdateCommentParams = zod.object({
  "id": zod.coerce.number()
})

export const updateCommentBodyContentMax = 500;



export const UpdateCommentBody = zod.object({
  "content": zod.string().max(updateCommentBodyContentMax)
})

export const UpdateCommentResponse = zod.object({
  "id": zod.number(),
  "postId": zod.number(),
  "authorId": zod.string(),
  "authorName": zod.string(),
  "authorImageUrl": zod.string().nullish(),
  "content": zod.string(),
  "createdAt": zod.coerce.date()
})


/**
 * @summary Delete a comment (owner only)
 */
export const DeleteCommentParams = zod.object({
  "id": zod.coerce.number()
})


/**
 * @summary Get current authenticated user info
 */
export const GetMeResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "imageUrl": zod.string().nullish(),
  "postCount": zod.number()
})


/**
 * @summary Get feed statistics (total posts, active users today)
 */
export const GetFeedStatsResponse = zod.object({
  "totalPosts": zod.number(),
  "totalComments": zod.number()
})


/**
 * @summary Upload a media asset for a post (owner only)
 */
export const UploadMediaBody = zod.object({
  "file": zod.instanceof(File)
})


/**
 * @summary Fetch an uploaded media asset
 */
export const GetMediaParams = zod.object({
  "fileName": zod.coerce.string()
})